refactor(notifications): move the overlay widgets out of the service
notification_service.dart was 496 lines and held two StatefulWidgets. It
both talked to the app and drew it, which CLAUDE.md separates: services
talk to the network, widgets render.

This is also why the design guard found twelve violations in a file I had
swept past. UI code was hiding somewhere nobody treats as UI, so neither
the module-by-module conversion nor a reader looking for widgets would
find it.

NotificationBanner and ToastMessage move to widgets/, public now that they
cross a file boundary. The service keeps showBanner and showToast, which
is the part that is genuinely a service concern -- inserting an
OverlayEntry and removing it again. 496 lines becomes 238.
